Move splitSearchString and stringsToRegExp to src/utils/string file
